+++ entity/src/stake_credential.rs
@@ -4,8 +4,8 @@ use serde::{Deserialize, Serialize};
 #[derive(Clone, Debug, PartialEq, DeriveEntityModel, Deserialize, Serialize)]
 #[sea_orm(table_name = "StakeCredential")]
 pub struct Model {
-    #[sea_orm(primary_key)]
-    pub id: i32,
+    #[sea_orm(primary_key, column_type = "BigInteger")]
+    pub id: i64,
     #[sea_orm(unique)]
     pub credential: Vec<u8>,
 }
